requirements.txt
psycopg2
sqlalchemy
rich
discord.py
fastapi[standard]